rename NotFound to 404
