+++ src/pages/404.js
@@ -2,7 +2,7 @@ import React, { useRef } from "react"
 
 import SEO from "../components/SEO/SEO"
 import Nav from "../components/Nav/Nav"
-import NotFound from "../components/NotFound/NotFound"
+import NotFound from "../components/404/404"
 import Footer from "../components/Footer/Footer"
 
 const scrollTo = ref => {
